Add files via upload
This table presents the topics identified by the authors of the paper. 

Below a summay of the comment is provided.
In the various degree courses, the presence of the topic of internationalization, multidisciplinarity and attention to information technology and innovation is noted. Additionally, all study programs appear to have a strong focus on practical applications of their field of study.
However, there are also significant differences between the various courses of study. For example, Law courses focus on legal systems and norms, while Humanities courses focus on literature, history and philosophy. Language courses focus on foreign languages and communication, while Mathematics courses focus on mathematics and its applications. Computer Science courses focus on computer science and its applications in business, while Engineering courses focus on engineering in various application fields.
Furthermore, some topics appear to be specific to some courses of study. For example, Humanities courses place a strong emphasis on the protection of cultural heritage, while Computer Science courses may include aspects related to information security and the social and economic aspects of technology. Finally, some courses of study may include specific linguistic knowledge, as in the case of courses in Law and Languages.